+++ src/qt/transactiondesc.cpp
@@ -20,7 +20,7 @@
 #include <policy/policy.h>
 #include <util/system.h>
 #include <validation.h>
-#include <wallet/ismine.h>
+#include <wallet/types.h>
 
 #include <stdint.h>
 #include <string>
